refactor: group src/app/components into domain subfolders
Flatten-out cleanup of the components/ dump into intent-named groups; rename the
vaguely-named cena/ to home/. No code changes, only moves + import updates.
- layout/   Navigation, Footer, AuthModal
- home/     HomePage + hero/rail (was cena/)
- study/    SpacedRepetitionGame, StudyCard, StudyProgressBar
- phrase/   AnkiExporter, PhraseCard, PhraseSortAndFilter, MetadataEditor/,
            DuplicatePhraseManager/, PhraseEditor/
- common/   Avatar, Breadcrumb, FavoriteButton, JobStatusBanner, ProfileStats,
            ProtectedRoute, UserDropdown
- ui/       (unchanged) ModalBase, FormField, StatCard
ShowSelector/ left at top level (self-contained, no clear group).

Verified: tsc clean, lint 0 errors (30 pre-existing warnings), build succeeds.
