+++ src/app/[series]/[episode]/edit/page.tsx
@@ -2,9 +2,9 @@
 
 import Link from "next/link";
 
-import MetadataEditor from "@/app/components/MetadataEditor";
-import PhraseEditor from "@/app/components/PhraseEditor";
-import { AdminRoute } from "@/app/components/ProtectedRoute";
+import MetadataEditor from "@/app/components/phrase/MetadataEditor";
+import PhraseEditor from "@/app/components/phrase/PhraseEditor";
+import { AdminRoute } from "@/app/components/common/ProtectedRoute";
 import { useEpisodeEdit } from "@/hooks/useEpisodeEdit";
 
 import EpisodeEditHeader from "./components/EpisodeEditHeader";
