+++ src/app/components/common/Avatar.tsx
