+++ src/app/components/common/FavoriteButton.tsx
