+++ src/app/components/common/JobStatusBanner.tsx
