+++ src/app/components/common/ProtectedRoute.tsx
