+++ src/app/components/phrase/AnkiExporter.tsx
