+++ src/app/components/phrase/DuplicatePhraseManager/BulkActionsBar.tsx
