+++ src/app/components/phrase/DuplicatePhraseManager/PhraseVariantRow.tsx
