+++ src/app/components/phrase/MetadataEditor/ShowMetadataForm.tsx
