+++ src/app/components/phrase/MetadataEditor/index.tsx
