+++ src/app/components/phrase/PhraseEditor/PhraseListItem.tsx
