+++ src/app/components/phrase/PhraseEditor/PhraseStatsBar.tsx
