+++ src/app/components/phrase/PhraseEditor/usePhraseEditor.ts
