+++ src/app/components/study/StudyProgressBar.tsx
