+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import HomePage from "./components/HomePage";
+import HomePage from "./components/home/HomePage";
 import { PhraseExtractionService, LibraryShow } from "@/lib/supabase";
 
 export default async function Home() {
